requirements.txt
altair==5.2.0
attrs==23.2.0
blinker==1.7.0
cachetools==5.3.3
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
coloredlogs==15.0.1
flatbuffers==24.3.7
gitdb==4.0.11
GitPython==3.1.42
humanfriendly==10.0
idna==3.6
imageio==2.34.0
Jinja2==3.1.3
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
lazy_loader==0.3
llvmlite==0.42.0
markdown-it-py==3.0.0
MarkupSafe==2.1.5
mdurl==0.1.2
mpmath==1.3.0
networkx==3.2.1
numba==0.59.0
numpy==1.26.4
onnxruntime==1.17.1
opencv-python-headless==4.9.0.80
packaging==23.2
pandas==2.2.1
pillow==10.2.0
platformdirs==4.2.0
pooch==1.8.1
protobuf==4.25.3
pyarrow==15.0.2
pydeck==0.8.1b0
Pygments==2.17.2
PyMatting==1.1.12
python-dateutil==2.9.0.post0
pytz==2024.1
referencing==0.34.0
rembg==2.0.56
requests==2.31.0
rich==13.7.1
rpds-py==0.18.0
scikit-image==0.22.0
scipy==1.12.0
six==1.16.0
smmap==5.0.1
streamlit==1.32.2
sympy==1.12
tenacity==8.2.3
tifffile==2024.2.12
toml==0.10.2
toolz==0.12.1
tornado==6.4
tqdm==4.66.2
typing_extensions==4.10.0
tzdata==2024.1
urllib3==2.2.1
watchdog==4.0.0
